+++ packages/sdk/src/types.ts
@@ -1,29 +1,29 @@
 /**
  * @module Common GRPC-Client
  */
-import type { Upward } from './grpc/upward.js';
-import { Cbor, TokenId } from './plt/index.js';
-import { TokenAccountInfo } from './plt/types.js';
-import * as AccountAddress from './types/AccountAddress.js';
+import type { Upward } from './grpc/index.js';
+import type { Cbor, TokenId } from './plt/index.js';
+import type { TokenAccountInfo } from './plt/types.js';
+import type * as AccountAddress from './types/AccountAddress.js';
 import type * as BlockHash from './types/BlockHash.js';
 import type * as CcdAmount from './types/CcdAmount.js';
-import * as ContractAddress from './types/ContractAddress.js';
+import type * as ContractAddress from './types/ContractAddress.js';
 import type * as ContractName from './types/ContractName.js';
-import * as CredentialRegistrationId from './types/CredentialRegistrationId.js';
-import { DataBlob } from './types/DataBlob.js';
-import * as Duration from './types/Duration.js';
-import * as Energy from './types/Energy.js';
+import type * as CredentialRegistrationId from './types/CredentialRegistrationId.js';
+import type { DataBlob } from './types/DataBlob.js';
+import type * as Duration from './types/Duration.js';
+import type * as Energy from './types/Energy.js';
 import type * as InitName from './types/InitName.js';
 import type * as ModuleReference from './types/ModuleReference.js';
-import * as Parameter from './types/Parameter.js';
+import type * as Parameter from './types/Parameter.js';
 import type * as ReceiveName from './types/ReceiveName.js';
 import type * as ReturnValue from './types/ReturnValue.js';
 import type * as SequenceNumber from './types/SequenceNumber.js';
-import * as Timestamp from './types/Timestamp.js';
+import type * as Timestamp from './types/Timestamp.js';
 import type * as TransactionExpiry from './types/TransactionExpiry.js';
 import type * as TransactionHash from './types/TransactionHash.js';
-import { RejectReason } from './types/rejectReason.js';
-import { ContractTraceEvent } from './types/transactionEvent.js';
+import type { RejectReason } from './types/rejectReason.js';
+import type { ContractTraceEvent } from './types/transactionEvent.js';
 
 export * from './types/NodeInfo.js';
 export * from './types/PeerInfo.js';
@@ -886,7 +886,13 @@ export type BakerId = bigint;
 export type DelegatorId = bigint;
 
 export interface BakerPoolInfo {
-    openStatus: OpenStatusText;
+    /**
+     * The status of validator pool
+     *
+     * **Please note**, this can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     */
+    openStatus: Upward<OpenStatusText>;
     metadataUrl: UrlString;
     commissionRates: CommissionRates;
 }
@@ -1103,6 +1109,7 @@ export enum AccountInfoType {
     Simple = 'simple',
     Baker = 'baker',
     Delegator = 'delegator',
+    Unknown = 'unknown',
 }
 
 interface AccountInfoCommon {
@@ -1171,7 +1178,20 @@ export interface AccountInfoDelegator extends AccountInfoCommon {
     accountDelegation: AccountDelegationDetails;
 }
 
-export type AccountInfo = AccountInfoSimple | AccountInfoBaker | AccountInfoDelegator;
+export interface AccountInfoUnknown extends AccountInfoCommon {
+    type: AccountInfoType.Unknown;
+    /**
+     * This will only ever be `null`, which represents a variant of staking info for the account which is
+     * unknown to the SDK, for known staking variants this is represented by either {@linkcode AccountInfoBaker}
+     * or {@linkcode AccountInfoDElegator}.
+     *
+     * **Note**: This field is named `accountBaker` to align with the JSON representation produced by the
+     * corresponding rust SDK.
+     */
+    accountBaker: Upward<never>;
+}
+
+export type AccountInfo = AccountInfoSimple | AccountInfoBaker | AccountInfoDelegator | AccountInfoUnknown;
 
 export interface Description {
     name: string;
@@ -2021,6 +2041,11 @@ export type Cooldown = {
     timestamp: Timestamp.Type;
     /** The amount that is in cooldown and set to be released at the end of the cooldown period */
     amount: CcdAmount.Type;
-    /** The status of the cooldown */
-    status: CooldownStatus;
+    /**
+     * The status of the cooldown
+     *
+     * **Please note**, this can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     */
+    status: Upward<CooldownStatus>;
 };